게시글 본인만 수정 및 삭제 가능하게 하기

+++ api_project/src/main/java/com/kch/api_project/controller/PostController.java
@@ -48,7 +48,7 @@ public class PostController {
             @PathVariable int id,
             @RequestBody PatachTestPostDTO dto
     ) {
-        postService.patachTestPost(id, dto);
+        postService.patachTestPost((long) id, dto);
         return ResponseEntity.ok(ApiResponse.ok("수정 성공"));
     }
